# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This Api is to handle dentistas paciente and turnos",
        "title": "Web Api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/dentistas": {
            "post": {
                "description": "Crea un nuevo dentista",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentista"
                ],
                "summary": "Crear dentista",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Información del dentista a crear",
                        "name": "dentista",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dentista"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/dentistas/{id}": {
            "get": {
                "description": "Obtiene un dentista por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentista"
                ],
                "summary": "Obtener dentista por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del dentista",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "description": "Actualiza los datos de un dentista existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentista"
                ],
                "summary": "Actualizar dentista por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del dentista a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del dentista a actualizar",
                        "name": "dentista",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dentista"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            },
            "delete": {
                "description": "Elimina un dentista por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentista"
                ],
                "summary": "Eliminar dentista por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del dentista a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente los datos de un dentista existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentista"
                ],
                "summary": "Actualizar parcialmente dentista por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del dentista a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar del dentista",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dentista"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/pacientes": {
            "post": {
                "description": "Crea un nuevo paciente con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Crear nuevo paciente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos del paciente",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/pacientes/{id}": {
            "get": {
                "description": "Obtiene los datos de un paciente existente por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Obtener paciente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del paciente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "description": "Actualiza un paciente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Actualizar paciente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del paciente a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del paciente a actualizar",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            },
            "delete": {
                "description": "Elimina un paciente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Eliminar paciente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del paciente a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un paciente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Actualizar parcialmente paciente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del paciente a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del paciente a actualizar",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/turnos": {
            "post": {
                "description": "Crea un nuevo turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Crear turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos del turno a crear",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/turnos/{id}": {
            "get": {
                "description": "Obtiene un turno por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Obtener turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a obtener",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "description": "Actualiza un turno existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Actualizar turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del turno a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del turno a actualizar",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            },
            "delete": {
                "description": "Elimina un turno existente por su ID",
                "tags": [
                    "Turno"
                ],
                "summary": "Eliminar turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del turno a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un turno existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Actualizar parcialmente un turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autenticación",
                        "name": "TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del turno a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos parciales del turno a actualizar",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Dentista": {
            "type": "object",
            "required": [
                "apellido",
                "matricula",
                "nombre"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "domain.Paciente": {
            "type": "object",
            "required": [
                "apellido",
                "dni",
                "fechaAlta",
                "nombre"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "dni": {
                    "type": "number"
                },
                "fechaAlta": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "domain.Turno": {
            "type": "object",
            "required": [
                "descripcion",
                "fechaHora",
                "odontologo",
                "paciente"
            ],
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "fechaHora": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "odontologo": {
                    "type": "string"
                },
                "paciente": {
                    "type": "string"
                }
            }
        }
    }
}